Update the code according to new HNB site Update
There is no way to get the HNB exchange rate earlier.
So we used to extract the data from web site using `cheerio`.
But after recent their web site update we have an API to
get the data and previous ex data from website doesn't work.

So through this PR removed the `cheerio` dependency and used the HNB
API to fetch the exchange rate.